Add modelstate condition to controllers

+++ NewsAggregator/NewsAggregator.App/Controllers/ArticleController.cs
@@ -87,12 +87,15 @@ namespace NewsAggregator.App.Controllers
         {
             try
             {
-                if (model != null)
+                if (ModelState.IsValid)
                 {
-                    await _articleService.CreateAsync(_mapper.Map<CreateOrEditArticleDto>(model));
-                    return RedirectToAction("GetArticlesOnAdminPanel", "Admin");
+                    if (model != null)
+                    {
+                        await _articleService.CreateAsync(_mapper.Map<CreateOrEditArticleDto>(model));
+                        return RedirectToAction("GetArticlesOnAdminPanel", "Admin");
+                    }
                 }
-                    
+    
                 return View(model);
             }
             catch (Exception ex)
@@ -138,12 +141,14 @@ namespace NewsAggregator.App.Controllers
         {
             try
             {
-                if (model != null)
+                if (ModelState.IsValid)
                 {
-                    await _articleService.UpdateAsync(_mapper.Map<CreateOrEditArticleDto>(model));
-                    return RedirectToAction("GetArticlesOnAdminPanel", "Admin");
+                    if (model != null)
+                    {
+                        await _articleService.UpdateAsync(_mapper.Map<CreateOrEditArticleDto>(model));
+                        return RedirectToAction("GetArticlesOnAdminPanel", "Admin");
+                    }
                 }
-
                 return View(model);    
             }
             catch (Exception ex)
@@ -177,13 +182,17 @@ namespace NewsAggregator.App.Controllers
         {
             try
             {
-                var delete = await _articleService.DeleteAsync(model.Id);
-                if (delete == null)
+                if (ModelState.IsValid)
                 {
-                    _logger.LogWarning($"{DateTime.Now}: Model is null in DeleteArticle method");
-                    return BadRequest();
+                    var delete = await _articleService.DeleteAsync(model.Id);
+                    if (delete == null)
+                    {
+                        _logger.LogWarning($"{DateTime.Now}: Model is null in DeleteArticle method");
+                        return BadRequest();
+                    }
+                    return RedirectToAction("GetArticlesOnAdminPanel", "Admin");
                 }
-                return RedirectToAction("GetArticlesOnAdminPanel", "Admin");
+                return View(model);
             }
             catch (Exception ex)
             {
